BSSLX-67: Fix after review

+++ odoo/songs/migration/post.py
@@ -47,27 +47,32 @@ def clean_modules_list(ctx):
 def clean_duplicated_menu(ctx):
     """ Clean duplicated menu """
     executive_summary_menu_to_delete = ctx.env.ref(
-        '__export__.ir_ui_menu_218'
+        '__export__.ir_ui_menu_218',
+        raise_if_not_found=False,
     )
-    executive_summary_menu_to_keep = ctx.env.ref(
-        'account_reports.account_financial_html_report_menu_4'
-    )
-    action_to_delete = executive_summary_menu_to_delete.action
-    action_to_keep = executive_summary_menu_to_keep.action
-    executive_summary_menu_to_keep.write({
-        'action':
-            '%s,%s' % (action_to_keep._name, action_to_keep.id),
-        'groups_id':
-            [(6, False, executive_summary_menu_to_delete.groups_id.ids)]
-    })
-    action_to_delete.unlink()
-    executive_summary_menu_to_delete.unlink()
+    if executive_summary_menu_to_delete:
+        executive_summary_menu_to_keep = ctx.env.ref(
+            'account_reports.account_financial_html_report_menu_4',
+        )
+        action_to_delete = executive_summary_menu_to_delete.action
+        action_to_keep = executive_summary_menu_to_keep.action
+        executive_summary_menu_to_keep.write({
+            'action':
+                '%s,%s' % (action_to_keep._name, action_to_keep.id),
+            'groups_id':
+                [(6, False, executive_summary_menu_to_delete.groups_id.ids)]
+        })
+        action_to_delete.unlink()
+        executive_summary_menu_to_delete.unlink()
 
 
 @anthem.log
 def migrate_account_reconcile_rule_values(ctx):
     """ Migrate account reconcile rule values """
-    ctx.env.cr.execute("""
+    if not ctx.env['account.reconcile.rule'].search([]):
+        # if account reconcile rule not found, we need
+        # to migrate the previous values in account_operation_rule table
+        ctx.env.cr.execute("""
 INSERT INTO
     account_reconcile_rule
 SELECT
@@ -92,7 +97,7 @@ SELECT
     account_operation_template_id AS account_reconcile_model_id
 FROM
     account_operation_rule_account_operation_template_rel;
-    """)
+        """)
 
 
 @anthem.log
